modif en long

+++ src/main/java/tp/calcul_pi/Master.java
@@ -16,7 +16,7 @@ import tp.calcul_pi.Pi;
 public class Master {
     public Master(){}
 
-    public long doRun(int totalCount, int numWorkers, String filename) throws InterruptedException, ExecutionException {
+    public long doRun(long totalCount, int numWorkers, String filename) throws InterruptedException, ExecutionException {
         boolean asFilename = !filename.isEmpty();
 
         long startTime = System.currentTimeMillis();
