+++ src/main/java/tp/calcul_pi/Worker.java
@@ -18,9 +18,9 @@ import tp.calcul_pi.Pi;
  * Task for running the Monte Carlo simulation.
  */
 class Worker implements Callable<Long> {
-    private int numIterations;
+    private long numIterations;
 
-    public Worker(int num) {
+    public Worker(long num) {
         this.numIterations = num;
     }
 
@@ -28,7 +28,7 @@ class Worker implements Callable<Long> {
     public Long call() {
         long circleCount = 0;
         Random prng = new Random();
-        for (int j = 0; j < numIterations; j++) {
+        for (long j = 0; j < numIterations; j++) {
             double x = prng.nextDouble();
             double y = prng.nextDouble();
             if ((x * x + y * y) < 1) ++circleCount;
